Support phased restarts with standalone exporter - fixes #22
This change adds some additional tests to cover clustered mode with the
standalone exporter. The phased restart example fails without the patch
to the exporter, which modifies it to handle the events differently:

 - The metrics server variable is moved outside of the on_booted handler
   so it is not created separately in the forked workers
 - The on_booted handler safely checks whether the server is running
   before attempting to start
 - The on_stopped and on_restart events are caught and issue a
   synchronous stop, rather than the previous state hook, which does not
   signal restart

The construction is a little odd, with the start_server lambda assigning
the metrics server instance, but I found the behavior quite fickle when
trying to reuse an instance. This was the most concise solution I could
come up with.